# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipe Tracker API",
        "description": "This is the API documentation for the Recipe Tracker project.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get all recipes",
                "operationId": "8b9e556d4d747bd60232961951494bfb",
                "responses": {
                    "200": {
                        "description": "List of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recipes fetched successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "recipes": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Recipe"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "operationId": "46448289940e329a644b891be9c66943",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "ingredients",
                                    "description",
                                    "prep_time",
                                    "cook_time",
                                    "difficulty"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Spaghetti Carbonara"
                                    },
                                    "ingredients": {
                                        "type": "string",
                                        "example": "Spaghetti, Eggs, Bacon"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Boil pasta. Fry bacon. Mix with eggs."
                                    },
                                    "prep_time": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "cook_time": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "enum": [
                                            "easy",
                                            "medium",
                                            "hard"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get a specific recipe by ID",
                "operationId": "494913cea16e9bd1ed0d3a25eda1b1f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe fetched successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "recipe": {
                                                    "$ref": "#/components/schemas/Recipe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Update an existing recipe",
                "operationId": "61665c36ed1115a4d53992f26035a8fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "ingredients": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "prep_time": {
                                        "type": "integer"
                                    },
                                    "cook_time": {
                                        "type": "integer"
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "enum": [
                                            "easy",
                                            "medium",
                                            "hard"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipe updated successfully"
                    },
                    "404": {
                        "description": "Recipe not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete a recipe",
                "operationId": "ff2f374ba56017e8d111c4e218868b9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recipe deleted successfully"
                    },
                    "404": {
                        "description": "Recipe not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recipes/search": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Search recipes by ingredients and time range",
                "operationId": "8c7faacbc706052cc061f04f19415189",
                "parameters": [
                    {
                        "name": "ingredients[]",
                        "in": "query",
                        "description": "List of ingredients",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "chicken"
                            }
                        }
                    },
                    {
                        "name": "min_time",
                        "in": "query",
                        "description": "Minimum total time (prep + cook)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "max_time",
                        "in": "query",
                        "description": "Maximum total time (prep + cook)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 60
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipes matching criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recipes fetched successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Recipe"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recipes/difficulty/{level}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Filter recipes by difficulty",
                "operationId": "787d9ea952676dcd086d6c5844ad0393",
                "parameters": [
                    {
                        "name": "level",
                        "in": "path",
                        "description": "Difficulty level (easy, medium, hard)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "easy",
                                "medium",
                                "hard"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recipes fetched successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Recipe"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid difficulty level"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Recipe": {
                "title": "Recipe",
                "description": "Recipe model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Spaghetti Bolognese"
                    },
                    "ingredients": {
                        "type": "string",
                        "example": "pasta, ground beef, tomatoes, onions, garlic, olive oil, salt, pepper"
                    },
                    "prep_time": {
                        "description": "Preparation time in minutes",
                        "type": "integer",
                        "example": 15
                    },
                    "cook_time": {
                        "description": "Cooking time in minutes",
                        "type": "integer",
                        "example": 30
                    },
                    "difficulty": {
                        "type": "string",
                        "enum": [
                            "easy",
                            "medium",
                            "hard"
                        ],
                        "example": "medium"
                    },
                    "description": {
                        "type": "string",
                        "example": "A classic Italian pasta dish made with ground beef and tomato sauce."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-10T15:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}